Users can now reset their password

+++ src/main/java/bibliotecame/back/Verification/VerificationController.java
@@ -8,11 +8,10 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/verify")
 public class VerificationController {
 
     private final UserService userService;
@@ -24,13 +23,30 @@ public class VerificationController {
         this.verificationService = verificationService;
     }
 
-    @PutMapping("/{token}")
+    @PutMapping("/verify/{token}")
     public ResponseEntity verifyAccount(@PathVariable String token){
         try{
             VerificationModel verificationModel = verificationService.findVerificationByToken(token);
             UserModel user = verificationModel.getUserModel();
             user.setVerified(true);
             userService.saveWithoutEncryption(user);
+            verificationService.deleteVerification(verificationModel);
+            return ResponseEntity.ok(user);
+        }catch (RuntimeException e){
+            return new ResponseEntity(new ErrorMessage("¡El token solicitado no es valido!"), HttpStatus.BAD_REQUEST);
+        }
+    }
+    @PutMapping("/reset/{token}")
+    public ResponseEntity resetPassword(@PathVariable String token, @RequestBody UserModel userModel){
+        try{
+            VerificationModel verificationModel = verificationService.findVerificationByToken(token);
+            if(verificationModel.getToken().length()<=40) return new ResponseEntity(new ErrorMessage("¡El token enviado no es un token de reinicio de contraseña valido!"),HttpStatus.BAD_REQUEST);
+            UserModel user = verificationModel.getUserModel();
+            if(user.getId()!=userModel.getId()) return new ResponseEntity(new ErrorMessage("¡Los datos recibidos son erroneos!"),HttpStatus.BAD_REQUEST);
+            String passwordRegex = "^(?=.*\\d)(?=.*[a-zA-Z])([a-zA-Z0-9]+){7,}$";
+            if(!userModel.getPassword().matches(passwordRegex)) return new ResponseEntity(new ErrorMessage("¡La contraseña ingresada no es valida, por favor intente con otra!"), HttpStatus.BAD_REQUEST);
+            user.setPassword(userModel.getPassword());
+            userService.saveUser(user);
             verificationService.deleteVerification(verificationModel);
             return ResponseEntity.ok(user);
         }catch (RuntimeException e){
